Move js/src/frontend-rs to js/src/frontend/smoosh (fixes #36)
