+++ .clang-format-ignore
@@ -22,6 +22,9 @@ js/src/irregexp/RegExpCharacters.cpp
 js/src/util/Unicode.cpp
 js/src/util/UnicodeNonBMP.h
 
+# Generated by cbindgen from js/src/frontend/smoosh/src/lib.rs
+js/src/frontend/smoosh/smoosh.h
+
 # Ignored because of bug 1506117 & 1342657
 layout/style/nsCSSAnonBoxList.h
 layout/style/nsCSSCounterDescList.h
