+++ Cargo.toml
@@ -7,7 +7,7 @@
 members = [
   "js/rust",
   "js/src",
-  "js/src/frontend-rs",
+  "js/src/frontend/smoosh",
   "js/src/frontend/binast", # Code generator.
   "js/src/rust",
   "js/src/wasm/cranelift",
