--- js/src/frontend-rs/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "frontend-rs"
-version = "0.1.0"
-authors = ["Ashley Hauck <khyperia@mozilla.com>"]
-edition = "2018"
-
-[dependencies]
-ast = { path = "../../jsparagus/rust/ast" }
-bumpalo = "2.6.0"
-emitter = { path = "../../jsparagus/rust/emitter" }
-parser = { path = "../../jsparagus/rust/parser" }
-
-[build-dependencies]
-cbindgen = "0.12"
