+++ js/src/frontend/smoosh/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "smoosh"
+version = "0.1.0"
+authors = ["Ashley Hauck <khyperia@mozilla.com>"]
+edition = "2018"
+
+[dependencies]
+ast = { path = "../../../jsparagus/rust/ast" }
+bumpalo = "2.6.0"
+emitter = { path = "../../../jsparagus/rust/emitter" }
+parser = { path = "../../../jsparagus/rust/parser" }
+
+[build-dependencies]
+cbindgen = "0.12"
