+++ js/src/frontend/smoosh/build.rs
@@ -9,5 +9,5 @@ fn main() {
         .with_crate(crate_dir)
         .generate()
         .expect("Unable to generate bindings")
-        .write_to_file("frontend-rs.h");
+        .write_to_file("smoosh.h");
 }
