+++ js/src/frontend/smoosh/moz.build
@@ -9,10 +9,10 @@ FINAL_LIBRARY = 'js'
 # Includes should be relative to parent path
 LOCAL_INCLUDES += [
     '!..',
-    '..'
+    '../..'
 ]
 
-include('../js-config.mozbuild')
-include('../js-cxxflags.mozbuild')
+include('../../js-config.mozbuild')
+include('../../js-cxxflags.mozbuild')
 
-DIRS += ['../rust']
+DIRS += ['../../rust']
