+++ js/src/frontend/smoosh/smoosh.h
