+++ js/src/rust/shared/Cargo.toml
@@ -12,7 +12,7 @@ path = "lib.rs"
 baldrdash = { path = "../../wasm/cranelift", optional = true }
 encoding_c = "0.9.5"
 encoding_c_mem = "0.2.4"
-frontend-rs = { path = "../../frontend-rs" }
+smoosh = { path = "../../frontend/smoosh" }
 mozilla-central-workspace-hack = { path = "../../../../build/workspace-hack" }
 mozglue-static = { path = "../../../../mozglue/static/rust" }
